Chapter 15: Authentication
added authentication to the website by forcing a login before being able to go to the dashboard or any other pages. Then also added a signing out functionality in the side navigation. Could have more sophisticated error handling but that can be improved upon later. Main concpetes are the middleware.ts,  advantage of employing Middleware for this task is that the protected routes will not even start rendering until the Middleware verifies the authentication, enhancing both the security and performance of your application. auth config and auth logic in code.